分析数据管理架构

+++ imetadata/business/metadata/base/job/c_dmFilePathInfoEx.py
@@ -25,7 +25,7 @@ class CDMFilePathInfoEx(CFileInfoEx):
     __ds_storage__ = None
     __ds_file_or_path__ = None
 
-    def __init__(self, file_type, file_name_with_full_path, root_path, storage_id, file_or_path_id, parent_id, owner_obj_id,
+    def __init__(self, file_type, file_name_with_full_path, storage_id, file_or_path_id, parent_id, owner_obj_id,
                  db_server_id):
         """
 
@@ -36,8 +36,14 @@ class CDMFilePathInfoEx(CFileInfoEx):
             如果为None, 则首先根据文件相对路径和storage_id, 查找数据库中登记的标识, 如果不存在, 则自行创建uuid;
             如果不为空, 则表明数据库中已经存储该文件标识
         """
-        super().__init__(file_type, file_name_with_full_path, root_path)
         self.__storage_id__ = storage_id
+
+        self.__ds_storage__ = CFactory().give_me_db(self.__db_server_id__).one_row(
+            'select dstid, dsttitle, dstunipath, dstotheroption from dm2_storage where dstid = :dstID',
+            {'dstid': self.__storage_id__})
+
+        root_path = self.__ds_storage__.value_by_name(0, 'dstunipath', '')
+        super().__init__(file_type, file_name_with_full_path, root_path)
         self.__my_id__ = file_or_path_id
         self.__db_server_id__ = db_server_id
         self.__parent_id__ = parent_id
@@ -51,12 +57,47 @@ class CDMFilePathInfoEx(CFileInfoEx):
         """
         pass
 
+    def db_init_object_of_directory_by_id(self, dir_id):
+        sql_update_path_object = '''
+            update dm2_storage_directory
+            set dsd_object_confirm = 0, dsd_object_id = null, dsd_object_type = null
+                , dsdscanfilestatus = 1, dsdscandirstatus = 1, dsd_directory_valid = -1
+            where dsdid = :dsdid
+            '''
+
+        CFactory().give_me_db(self.__db_server_id__).execute(sql_update_path_object, {'dsdid': dir_id})
+
     def db_delete_object_by_id(self, object_id):
+        if object_id == '' or object_id is None:
+            return
+
         sql_delete_object_by_id = '''
         delete from dm2_storage_object
         where dsoid = :dsoID
         '''
-        CFactory.give_me_db(self.__db_server_id__).execute(sql_delete_object_by_id, {'dsoid': object_id})
+
+        sql_delete_object_details_by_id = '''
+        delete from dm2_storage_obj_detail
+        where dodid = :dsoID
+        '''
+
+        sql_delete_object_spatial_by_id = '''
+        delete from dm2_storage_object_spatial
+        where dsos_id = :dsoID
+        '''
+
+        engine = CFactory().give_me_db(self.__db_server_id__)
+        session = engine.give_me_session()
+        try:
+            engine.session_execute(session, sql_delete_object_details_by_id, {'dsoid': object_id})
+            engine.session_execute(session, sql_delete_object_spatial_by_id, {'dsoid': object_id})
+            engine.session_execute(session, sql_delete_object_by_id, {'dsoid': object_id})
+            engine.session_commit(session)
+        except Exception as error:
+            CLogger().warning('数据库处理出现异常, 错误信息为: {0}'.format(error.__str__))
+            engine.session_rollback(session)
+        finally:
+            engine.session_close(session)
 
     def white_black_valid(self):
         """
@@ -120,9 +161,9 @@ class CDMFilePathInfoEx(CFileInfoEx):
                 CLogger().debug(
                     '{0} is plugins_classified as {1}.{2}'.format(target, class_classified_obj.get_group_name(),
                                                                   class_classified_obj.get_id()))
-                return class_classified_obj
+                return class_classified_obj, object_confirm, object_name
         else:
-            return None
+            return None, self.Object_Confirm_IUnKnown, None
 
     def plugins(self, plugins_id: str) -> CPlugins:
         target = self.__file_main_name__
